+++ Front/src/pages/Admin/Users/User/User.js
@@ -1,18 +1,26 @@
 import { useTranslation } from "react-i18next";
 import { deleteUser } from "../../../../apis/Users";
 import style from "./User.module.scss";
+import { useContext } from "react";
+import { AdminContext } from "../../../../context/AdminContext";
 
 function User({ user, setIdUserEdit, users, setUsers }) {
   const { t } = useTranslation();
+  const {setShowDelete, deleteAlertResponse, setDeleteAlertResponse} = useContext(AdminContext);
 
   function onEdit() {
     setIdUserEdit(user.id);
   }
 
-  function onDelete() {
+  function onShowDeleteMessage(){
+    setShowDelete(true);
+  }
+
+  if(deleteAlertResponse){
     if (deleteUser(user.id)) {
       setUsers(users.filter((u) => u.id !== user.id));
     }
+    setDeleteAlertResponse(false)
   }
 
   return (
@@ -32,12 +40,15 @@ function User({ user, setIdUserEdit, users, setUsers }) {
       >
         {user.mail}
       </div>
+      <div className="d-flex justify-content-center align-items-center">
       <div className={`${style.edit} mr10`} onClick={onEdit}>
         {t("edit")}
       </div>
-      <div className={`${style.delete}`} onClick={onDelete}>
+      <div className={`${style.delete}`} onClick={onShowDeleteMessage}>
         {t("delete")}
       </div>
+      </div>
+      
     </div>
   );
 }
